server: apollo client setup
