+++ apps/server/package.json
@@ -8,12 +8,14 @@
     "seed": "ts-node src/seed.ts"
   },
   "dependencies": {
+    "@apollo/server": "^4.9.2",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
     "@types/mongoose": "^5.11.97",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-graphql": "^0.12.0",
+    "fs": "^0.0.1-security",
     "graphql": "^16.8.0",
     "mongoose": "^7.4.3",
     "typescript": "^5.1.6"
